Revert "create qr code"
