+++ Project-Mug/src/frontend/package.json
@@ -12,7 +12,6 @@
     "core-js": "^3.6.5",
     "vue": "^2.6.11",
     "vue-apexcharts": "^1.6.1",
-    "vue-qrcode": "^0.4.0",
     "vue-qrcode-reader": "^3.0.3",
     "vue-router": "^3.5.1"
   },
